requirements.txt


fastapi==0.68.1
tensorflow-cpu==2.8.0
uvicorn==0.15.0
Pillow==7.1.2
python-multipart==0.0.5

